Add support for Turtle data sources
