+++ nemo/src/io.rs
@@ -6,9 +6,9 @@ use std::io::Write;
 
 pub mod dsv;
 pub mod input_manager;
-pub mod ntriples;
 pub mod output_file_manager;
 pub mod parser;
+pub mod rdf_triples;
 
 pub use output_file_manager::OutputFileManager;
