+++ resources/testcases/rdf/ntriples-example2.rls
