+++ resources/testcases/rdf/ntriples-example2/goal.csv
