+++ resources/testcases/rdf/ntriples-example3.rls
