+++ resources/testcases/rdf/ntriples-example3/atoms.csv
