+++ resources/testcases/rdf/ntriples-example3/labels.csv
