+++ resources/testcases/rdf/sources/example2.nt
