+++ resources/testcases/rdf/sources/example3.nt
