+++ resources/testcases/rdf/turtle-example2/goal.csv
@@ -1,0 +1,1 @@
+42
